Moving serialization configuration to a common reachable project
