+++ Source/Kernel/Server/Program.cs
@@ -3,6 +3,7 @@
 
 using System.Globalization;
 using Aksio.Applications.Autofac;
+using Aksio.Cratis.Kernel.Serialization;
 using Serilog;
 
 #pragma warning disable SA1600
