+++ Source/Kernel/Server/Server.csproj
@@ -34,6 +34,7 @@
         <ProjectReference Include="../Persistence/Persistence.csproj" />
         <ProjectReference Include="../Reactions/Reactions.csproj" />
         <ProjectReference Include="../Read/Read.csproj" />
+        <ProjectReference Include="../Serialization/Serialization.csproj" />
         <ProjectReference Include="../Shared/Shared.csproj" />
     </ItemGroup>
