+++ cratis.sln
@@ -85,6 +85,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kernel", "Specifications\In
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Roslyn.Extensions", "Source\Roslyn.Extensions\Roslyn.Extensions.csproj", "{4B9A23BE-CE65-4A84-8BC7-819B359DF779}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Serialization", "Source\Kernel\Serialization\Serialization.csproj", "{42AEFC6B-816E-4474-B21D-5A7EBFC5539D}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -230,6 +232,10 @@ Global
 		{4B9A23BE-CE65-4A84-8BC7-819B359DF779}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4B9A23BE-CE65-4A84-8BC7-819B359DF779}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4B9A23BE-CE65-4A84-8BC7-819B359DF779}.Release|Any CPU.Build.0 = Release|Any CPU
+		{42AEFC6B-816E-4474-B21D-5A7EBFC5539D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{42AEFC6B-816E-4474-B21D-5A7EBFC5539D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{42AEFC6B-816E-4474-B21D-5A7EBFC5539D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{42AEFC6B-816E-4474-B21D-5A7EBFC5539D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{F1A70D86-CE4A-443C-B4F9-EEFBC3E896F3} = {216800F4-7820-46A5-A3AD-4FCE82927B8D}
@@ -260,5 +266,6 @@ Global
 		{540799D7-A383-4DCD-B7F4-96FD4056D7A1} = {7AD8BE82-9E08-46FB-ADCB-7D788BAFD76D}
 		{1C757327-F744-41EB-8489-D8C34FB4B1AE} = {25011E0A-4967-4A8E-BCC7-48FE8F8453C4}
 		{4B9A23BE-CE65-4A84-8BC7-819B359DF779} = {216800F4-7820-46A5-A3AD-4FCE82927B8D}
+		{42AEFC6B-816E-4474-B21D-5A7EBFC5539D} = {7AD8BE82-9E08-46FB-ADCB-7D788BAFD76D}
 	EndGlobalSection
 EndGlobal
